requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
ape-alchemy==0.2.0
ape-ens==0.2.0
ape-hardhat==0.1.0b3
ape-solidity==0.1.0b2
ape-tokens==0.1.0b4
appnope==0.1.2
APScheduler==3.6.3
argcomplete==2.0.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==1.0.3
bitarray==1.2.2
black==22.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.28
cytoolz==0.11.2
dataclassy==0.10.4
DateTime==4.4
decorator==5.1.1
Deprecated==1.2.13
discord.py @ git+https://github.com/Rapptz/discord.py@ae51d5849a946699362372fb9d7f637bb81186f3
distlib==0.3.4
eth-abi==2.1.1
eth-account==0.5.7
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b6
eth-typing==2.3.0
eth-utils==1.10.0
ethpm-types==0.3.1
evm-trace==0.1.0a3
executing==0.8.2
filelock==3.4.2
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
flatbuffers==1.12
fonttools==4.31.1
frozenlist==1.3.0
gast==0.4.0
google-auth==2.6.2
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hexbytes==0.2.2
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
langdetect==1.0.9
lru-dict==1.1.7
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
morphys==1.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
nltk==3.7
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.4.2
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipx==1.0.0
platformdirs==2.4.1
pluggy==0.13.1
praw==7.5.0
prawcore==2.3.0
prompt-toolkit==3.0.28
protobuf==3.19.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cid==0.3.0
py-cpuinfo==5.0.0
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.8.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.13.0
pydantic==1.9.0
pyethash==0.1.27
pygit2==1.8.0
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
pyrsistent==0.18.1
pysha3==1.0.2
pytest==6.2.5
python-baseconv==1.2.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rich==10.16.2
rlp==2.0.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==0.24.2
scipy==1.8.0
semantic-version==2.8.5
sentencepiece==0.1.96
six==1.15.0
sortedcontainers==2.4.0
squarify==0.4.3
stack-data==0.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
tk==0.1.0
tokenizers==0.0.11
tokenlists==0.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.17.0
trie==2.0.0a5
tweepy==4.7.0
tweet-preprocessor==0.6.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
update-checker==0.18.0
urllib3==1.26.8
userpath==1.7.0
varint==1.0.2
virtualenv==20.13.0
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.29.1
websocket-client==1.3.1
websockets==9.1
Werkzeug==2.0.3
wrapt==1.12.1
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
